completed script; db kinda works, but doesn't -- have to add SoldItem somehow to get sums

+++ POS/src/ee/ut/math/tvt/salessystem/domain/data/DisplayableItem.java
@@ -7,5 +7,5 @@ public interface DisplayableItem {
 	/**
 	 * Id of entity.
 	 */
-	public Long getId();
+	public Integer getId();
 }
